Packages' sets

These are presets of packages which can be used to easily prepare systems installing them in groups or all together (using `pkg_all`).

pkg_misc

pkg_utils

pkg_speed

pkg_mlt

pkg_stat

pkg_stan

pkg_devel

pkg_docs

pkg_prod

pkg_all

Format

An object of class character of length 3.

An object of class character of length 4.

An object of class character of length 4.

An object of class character of length 6.

An object of class character of length 7.

An object of class character of length 2.

An object of class character of length 12.

An object of class character of length 4.

An object of class character of length 4.

An object of class character of length 44.

Details

Packages imported by the mentioned packages are not included into the sets (that is the reason why, e.g., you do not see `shiny` or `knitr`, ...).

Functions

pkg_misc: Packages for side-interest purposes, including `beepr`, `fortunes`, and `telegram.bot`.

pkg_utils: Utilities for data analyses, including `here`, `lobstr`, `progress`, and `tidyverse`.

pkg_speed: Are you in a hurry?! This includes `furrr`, `parallel`, `snow`, and `tidyfast`.

pkg_mlt: Machine Learning in R (we do not include `keras` nor `tensorflow` which require _ad hoc_ configuration). This includes `caret`, `glmnet`, `mlr3`, `snow`, `tidymodels`, `tidytext`, and `tm`.

pkg_stat: (Clinical oriented) statistical analyses in R. This includes `CBPS`, `cobalt`, `MatchIt`, `mice`, `rms`, `twang`, and `WeightIt`.

pkg_stan: Stan analyses in R (Alert: these packages require `stan` be installed into the system. They are not included into `pkg_all`). This includes `rstanarm`, and `brms`.

pkg_devel: Develop stuff in R. This includes `assertive`, `covr`, `lintr`, `fs`, `profvis`, `pryr`, `renv`, `roxygen2`, `styler`, `testthat`, `spelling`, and `usethis`.

pkg_docs: Write and render things in R. This includes `blogdown`, `DT`, `pander`, and `pkgdown`.

pkg_prod: for develop and put in production stuffs (`shiny` app included). This includes `golem`, `docopt`, `plumbr`, and `shinyjs`.

pkg_all: Union of the `pkg_misc`, `pkg_utils`, `pkg_speed`, `pkg_mlt`, `pkg_stat`, `pkg_devel`, `pkg_docs`, and `pkg_prod` pkg_* sets.

See also