tidy_summary

Converts a summary() object produced by Hmisc or by rms packages to a tidy data frame ready to be pandered (e.g. printed on a word document after knit the source.

tidy_summary(x, ...)

# S3 method for summary.formula.reverse
tidy_summary(x, ...)

# S3 method for summary.rms
tidy_summary(x, diff_digits = 2, ...)

Arguments

x
an object used to select a method, output of some summary by Hmisc.

...
further arguments passed to or from other methods

diff_digits
number of significant digits to use (default 2) for the step-difference between continuous variable HR computation.

Value

a [tibble][tibble::tibble-package]

Methods (by class)

summary.formula.reverse: Tidies a summary reverse output from the summary.formula called with method = "reverse".

summary.rms: Convert the output of the summary.rms into a data frame, reporting only the Hazard Ratio with the .95 CI and the incremental step (for continuous variables) reference (for categorical variables) for which the Hazard is referred to (i.e. without \(\beta\)s, Low, High, S.E. and Type).

Note

to see the options you can pass to ... for a custom print, see the print section in summary.formula.

Examples

library(Hmisc)
my_summary <- summary(Species ~ ., data = iris, method = "reverse")
tidy_summary(my_summary)

#> # A tibble: 4 x 4
#>   `&nbsp;`     `setosa \n(N=50)`   `versicolor \n(N=50)` `virginica \n(N=50)`
#>   <chr>        <chr>               <chr>                 <chr>               
#> 1 Sepal.Length "4.800/5.000/5.200" "5.600/5.900/6.300"   "6.225/6.500/6.900" 
#> 2 Sepal.Width  "3.200/3.400/3.675" "2.525/2.800/3.000"   "2.800/3.000/3.175" 
#> 3 Petal.Length "1.400/1.500/1.575" "4.000/4.350/4.600"   "5.100/5.550/5.875" 
#> 4 Petal.Width  "   0.2/0.2/0.3"    "   1.2/1.3/1.5"      "   1.8/2.0/2.3"    

library(rms)
options(dd = "datadist")
n <- 1000
set.seed(731)
age <- 50 + 12 * rnorm(n)
sex <- factor(sample(c("Male", "Female"), n,
  rep = TRUE,
  prob = c(.6, .4)
))
cens <- 15 * runif(n)
h <- .02 * exp(.04 * (age - 50) + .8 * (sex == "Female"))
dt <- -log(runif(n)) / h
e <- ifelse(dt <= cens, 1, 0)
dt <- pmin(dt, cens)

dd <- datadist(age, sex)

S <- survival::Surv(dt, e)
f <- rms::cph(S ~ age + sex)

#> Error in Design(eval.parent(m)): dataset dd not found for options(datadist=)

my_summary <- summary(f)

#> Error in summary(f): object 'f' not found

tidy_summary(my_summary)

#> # A tibble: 4 x 4
#>   `&nbsp;`     `setosa \n(N=50)`   `versicolor \n(N=50)` `virginica \n(N=50)`
#>   <chr>        <chr>               <chr>                 <chr>               
#> 1 Sepal.Length "4.800/5.000/5.200" "5.600/5.900/6.300"   "6.225/6.500/6.900" 
#> 2 Sepal.Width  "3.200/3.400/3.675" "2.525/2.800/3.000"   "2.800/3.000/3.175" 
#> 3 Petal.Length "1.400/1.500/1.575" "4.000/4.350/4.600"   "5.100/5.550/5.875" 
#> 4 Petal.Width  "   0.2/0.2/0.3"    "   1.2/1.3/1.5"      "   1.8/2.0/2.3"